// frontend/src/api/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: string;
};

export type CreateJobDto = {
    /**
     * URL list for checking
     */
    urls: Array<string>;
};

export type CreateJobResponseDto = {
    jobId: string;
};

export type JobStatsDto = {
    total: number;
    processed: number;
    success: number;
    error: number;
};

export type JobListItemDto = {
    id: string;
    createdAt: string;
    status: 'pending' | 'in_progress' | 'completed' | 'cancelled' | 'failed';
    stats: JobStatsDto;
};

export type UrlCheckResultDto = {
    url: string;
    status: 'pending' | 'in_progress' | 'success' | 'error' | 'cancelled';
    httpStatus?: number;
    errorMessage?: string;
    startedAt?: string;
    finishedAt?: string;
    durationMs?: number;
};

export type JobDto = {
    id: string;
    createdAt: string;
    status: 'pending' | 'in_progress' | 'completed' | 'cancelled' | 'failed';
    stats: JobStatsDto;
    urls: Array<UrlCheckResultDto>;
};

export type JobsControllerFindAllData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/jobs';
};

export type JobsControllerFindAllResponses = {
    200: Array<JobListItemDto>;
};

export type JobsControllerFindAllResponse = JobsControllerFindAllResponses[keyof JobsControllerFindAllResponses];

export type JobsControllerCreateData = {
    body: CreateJobDto;
    path?: never;
    query?: never;
    url: '/api/jobs';
};

export type JobsControllerCreateResponses = {
    201: CreateJobResponseDto;
};

export type JobsControllerCreateResponse = JobsControllerCreateResponses[keyof JobsControllerCreateResponses];

export type JobsControllerCancelData = {
    body?: never;
    path: {
        id: string;
    };
    query?: never;
    url: '/api/jobs/{id}';
};

export type JobsControllerCancelResponses = {
    200: JobDto;
};

export type JobsControllerCancelResponse = JobsControllerCancelResponses[keyof JobsControllerCancelResponses];

export type JobsControllerFindOneData = {
    body?: never;
    path: {
        id: string;
    };
    query?: never;
    url: '/api/jobs/{id}';
};

export type JobsControllerFindOneResponses = {
    200: JobDto;
};

export type JobsControllerFindOneResponse = JobsControllerFindOneResponses[keyof JobsControllerFindOneResponses];